Allow to specify multiple projects
Multiple projects can now be specified within one repository. All of
the commands besides "pull" require the --project flag to be provided
with project name specified in the .wti file. It is also possible to
set default project for pull command.

The .wti file should be structured as:
project_1:
  api_key: xxxx
project_2:
  api_key: xxxx
default:
  api_key: xxxx